change: funding banner
