+++ lib/main.ts
@@ -23,6 +23,7 @@ export { FooterLogoLinks } from './components/Footer/FooterLogoLinks/FooterLogoL
 export { FooterStandard } from './components/Footer/FooterStandard/FooterStandard'
 export { Forbidden403 } from './components/ErrorMessages/Forbidden403/Forbidden403'
 export { FundingDetails } from './components/FundingDetails/FundingDetails'
+export { ProgressBar } from './components/ProgressBar/ProgressBar'
 export { EmbedHubSpot } from './components/Embed/EmbedHubSpot.js'
 export { FullBanner } from './components/FullBanner/FullBanner'
 export { Figure } from './components/Figure/Figure'
